Upgraded to support PyJWT 0.4.0 which requires cryptograph
cryptography is a Python replacement for PyCrypto which hasn’t
been updated in over a year.  It uses OpenSSL directly and is
much faster - and has more bug fixes - is better supported.

As pyramid_jwtauth depends on PyJWT is seems better to track
PyJWT requirements rather than stick with PyCrypto.